Adds resource quota support for container tests
Enables setting resource limits (CPU, memory) for n8n main and
worker containers in test environments.

This change introduces environment variables to configure resource
quotas and applies them when creating n8n containers during
testing. This allows simulating resource-constrained environments
to reproduce and fix race conditions.
